+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-core/src/test/java/org/activiti/cloud/services/core/SecurityPoliciesApplicationServiceTest.java
@@ -1,7 +1,7 @@
 package org.activiti.cloud.services.core;
 
-import org.activiti.cloud.services.SecurityPolicy;
-import org.activiti.cloud.services.SecurityPoliciesService;
+import org.activiti.cloud.services.security.SecurityPolicy;
+import org.activiti.cloud.services.security.SecurityPoliciesService;
 import org.activiti.engine.UserGroupLookupProxy;
 import org.activiti.engine.UserRoleLookupProxy;
 import org.activiti.engine.repository.ProcessDefinitionQuery;
